Basically, I followed the guide that GitHub included in writing Markdown documents, like using "#" and "##" to create different levels of headers. Similarly, by studying the examples on GitHub, I used "*...*" and "**...**" to emphasize certain words/phrases. I used "*" to create lists, and typed in sub-points by placing two spaces before the "*". I also made use of the Emoji Cheat Sheet on the website, which enabled me to experiment with certain emoji, making my self-introduction more interesting.
The part about embedding an image was a bit more complicated, as I needed to go to my GitHub repo page first, click on "add file," and then copy the link to my readme file. Also, one concern I had was regarding "checking my process" before submitting the homework — the guide provided a comparison between the syntax and the actual format that we see (like the bolded text), but it seems that I could only see the syntax/code in RStudio Workbench, so I was not sure if I was doing everything correctly.

Overall, I think the process was relatively easy, as I could just follow the detailed instructions online and "fill in the blanks."